moved event models to their own folder
